My team is too good for me
