--- app/src/main/java/com/lastaoutdoor/lasta/data/api/dummyFIle.md
@@ -1,1 +1,0 @@
-Just added this file temporarily to be able to push it
