+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/Node.kt
@@ -1,4 +1,4 @@
-package com.lastaoutdoor.lasta.model.data
+package com.lastaoutdoor.lasta.data.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
