+++ app/src/main/java/com/lastaoutdoor/lasta/viewmodel/OutdoorActivityViewModel.kt
@@ -2,7 +2,7 @@ package com.lastaoutdoor.lasta.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
-import com.lastaoutdoor.lasta.model.repository.OutdoorActivityRepository
+import com.lastaoutdoor.lasta.repository.OutdoorActivityRepository
 
 class OutdoorActivityViewModel(application: Application) : AndroidViewModel(application) {
   private val repository: OutdoorActivityRepository by lazy {
